test(editor): move test util to helper
