+++ packages/form-js-editor/test/spec/properties-panel/components/NumberInputEntry.spec.js
@@ -6,7 +6,7 @@ import {
 
 import { NumberInputEntry } from '../../../../src/render/components/properties-panel/components';
 
-import { WithFormEditorContext } from './Util';
+import { WithFormEditorContext } from '../helper';
 
 import { insertStyles } from '../../../TestHelper';
