+++ packages/form-js-editor/test/spec/properties-panel/components/TextInput.spec.js
@@ -6,7 +6,7 @@ import {
 
 import { TextInput } from '../../../../src/render/components/properties-panel/components';
 
-import { WithFormEditorContext } from './Util';
+import { WithFormEditorContext } from '../helper';
 
 import { insertStyles } from '../../../TestHelper';
