+++ packages/form-js-editor/test/spec/properties-panel/components/Textarea.spec.js
@@ -6,7 +6,7 @@ import {
 
 import { Textarea } from '../../../../src/render/components/properties-panel/components';
 
-import { WithFormEditorContext } from './Util';
+import { WithFormEditorContext } from '../helper';
 
 import { insertStyles } from '../../../TestHelper';
